+++ packages/core/package.json
@@ -4,44 +4,72 @@
   "repository": "git@github.com:ateam-dev/react-portable.git",
   "license": "Apache-2.0",
   "bin": {
-    "react-portable": "./bin/cli.js"
+    "portable": "./dist/cli.cjs"
   },
-  "types": "./types/index.d.ts",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.mjs",
+      "node": "./dist/index.cjs",
+      "default": "./dist/index.mjs"
+    },
+    "./vite": {
+      "require": "./dist/vite.cjs",
+      "import": "./dist/vite.mjs",
+      "node": "./dist/vite.cjs",
+      "default": "./dist/vite.cjs"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      "vite": [
+        "./dist/vite.d.ts"
+      ]
+    }
+  },
   "files": [
-    "src",
-    "bin",
-    "types"
+    "dist",
+    "src/templates"
   ],
   "scripts": {
-    "build": "tsc -p tsconfig.cli.json && chmod a+x bin/cli.js",
-    "build:watch": "tsc -p tsconfig.cli.json --watch",
-    "typecheck": "tsc --noEmit -p tsconfig.cli.json"
+    "build": "unbuild",
+    "watch": "nodemon --ext ts,tsx --watch src -x 'npm run build'",
+    "typecheck": "tsc --noEmit",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage"
   },
   "devDependencies": {
     "@builder.io/qwik-react": "^0.5.0",
     "@cloudflare/workers-types": "^4.20230518.0",
-    "@types/fs-extra": "^11.0.1",
+    "@react-portable/client": "0.2.0",
+    "@testing-library/react": "^14.0.0",
     "@types/node": "^20.3.0",
+    "@vitest/coverage-v8": "^0.33.0",
+    "memfs": "^4.2.0",
+    "nodemon": "^2.0.22",
     "typescript": "5.1.3",
+    "unbuild": "^1.2.1",
+    "vitest": "^0.33.0",
     "wrangler": "^3.1.0"
   },
   "peerDependencies": {
     "@builder.io/qwik-react": ">=0.5.0",
     "react": ">=18.0.1",
     "react-dom": ">=18.0.0",
+    "typescript": ">=4",
+    "vite": ">=4",
     "wrangler": ">=3"
   },
   "dependencies": {
     "@builder.io/qwik": "^1.1.5",
     "@builder.io/qwik-city": "^1.1.5",
     "@hono/qwik-city": "^0.0.2",
-    "chokidar": "^3.5.3",
+    "app-root-path": "^3.1.0",
     "commander": "^10.0.1",
-    "fs-extra": "^11.1.1",
-    "glob": "^10.2.7",
-    "hono": "^3.2.5",
-    "vite": "^4.3.8",
-    "vite-tsconfig-paths": "4.2.0"
+    "hono": "^3.2.5"
   },
   "engines": {
     "node": ">=15.0.0"
