+++ packages/core/src/index.ts
@@ -1,0 +1,2 @@
+export { portable } from "./portable";
+export type { Strategy, PortableComponent, Loader } from "./portable";
