--- packages/core/src/react-portable.config.js
@@ -1,6 +1,0 @@
-const path = require("node:path");
-
-/** @type {import('@react-portable/core/types').Config} */
-module.exports = {
-  src: path.resolve(__dirname, "./src"),
-};
