+++ packages/core/src/templates/root.tsx
@@ -1,3 +1,4 @@
+/** @jsxImportSource @builder.io/qwik */
 import { QwikCityProvider, RouterOutlet } from "@builder.io/qwik-city";
 import { component$ } from "@builder.io/qwik";
 export default component$(() => (
