+++ packages/core/src/templates/route.tsx
@@ -1,9 +1,17 @@
+/** @jsxImportSource @builder.io/qwik */
 import { qwikify$ } from "@builder.io/qwik-react";
 import { component$ } from "@builder.io/qwik";
 import { RequestHandler, routeLoader$ } from "@builder.io/qwik-city";
-import * as Entry from "react-portable:virtual";
+import { PortableComponent } from "../portable";
+// @ts-ignore
+import * as Entries from "__entryPath__";
 
-const { hydrate, revalidate = 0 } = Entry.strategy ?? {};
+const Entry = Object.values(Entries).find((module) => {
+  if (typeof module === "function" && "__code" in module)
+    return module.__code === "__code__";
+}) as PortableComponent;
+
+const { hydrate = "onIdle", revalidate = 0 } = Entry.__strategy ?? {};
 
 const qwikifyOption =
   hydrate === "onUse"
@@ -20,10 +28,12 @@ const qwikifyOption =
         eagerness: undefined,
         event: undefined,
       };
-const QComponent = qwikify$(Entry.default, qwikifyOption);
+const QComponent__code__ = qwikify$(Entry, qwikifyOption);
 const getProps = routeLoader$(async ({ request, error }) => {
   try {
-    const res = (await Entry.loader?.(request, { error })) ?? {};
+    if (request.method === "POST") return await request.json();
+
+    const res = (await Entry.__loader?.(request, { error })) ?? {};
     if (res instanceof Error) throw res;
     return res;
   } catch (e) {
@@ -32,7 +42,7 @@ const getProps = routeLoader$(async ({ request, error }) => {
 });
 export default component$(() => {
   const props = getProps().value;
-  return <QComponent {...props} />;
+  return <QComponent__code__ {...props} />;
 });
 
 export const onRequest: RequestHandler = async (requestEvent) => {
