+++ packages/core/src/utils.ts
@@ -1,0 +1,1 @@
+export const currentDir = () => __dirname;
