--- packages/core/src/wrangler.toml
@@ -1,6 +1,0 @@
-name = "react-portable"
-main = ".rp/server/worker.mjs"
-compatibility_date = "2023-05-10"
-
-[site]
-bucket = ".rp/client"
