--- packages/core/tsconfig.cli.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2015",
-    "module": "CommonJS",
-    "strict": true,
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "skipLibCheck": true,
-    "outDir": "bin"
-  },
-  "include": [
-    "cli.ts"
-  ],
-}
