--- packages/core/tsconfig.qwik.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "target": "ES2017",
-    "module": "ES2020",
-    "lib": ["es2020", "DOM", "WebWorker", "DOM.Iterable"],
-    "jsx": "react-jsx",
-    "jsxImportSource": "@builder.io/qwik",
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "isolatedModules": true,
-    "types": ["node", "vite/client", "@cloudflare/workers-types"],
-    "outDir": "dist"
-  },
-  "include": [
-    "src", "types"
-  ]
-}
