--- packages/core/types/index.d.ts
@@ -1,3 +1,0 @@
-import "./module";
-export { Strategy, Loader } from "react-portable:virtual";
-export { Config } from "./config";
